+++ app/src/main/AndroidManifest.xml
@@ -27,7 +27,7 @@
             android:name=".ui.profile.ProfileActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.AuthenticationActivity"
+            android:name=".ui.auth.AuthenticationActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
