+++ app/src/main/java/com/bangkit/alpaca/ui/auth/forgotpassword/ForgotPasswordViewModel.kt
@@ -1,4 +1,4 @@
-package com.bangkit.alpaca.ui.forgotpassword
+package com.bangkit.alpaca.ui.auth.forgotpassword
 
 import androidx.lifecycle.ViewModel
