+++ app/src/main/java/com/bangkit/alpaca/ui/auth/login/LoginViewModel.kt
@@ -1,4 +1,4 @@
-package com.bangkit.alpaca.ui.login
+package com.bangkit.alpaca.ui.auth.login
 
 import androidx.lifecycle.ViewModel
