+++ app/src/main/res/layout/activity_authentication.xml
@@ -5,7 +5,7 @@
     android:id="@+id/auth_container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.AuthenticationActivity">
+    tools:context=".ui.auth.AuthenticationActivity">
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/nav_host_fragment_activity_auth"
