+++ app/src/main/res/layout/fragment_forgot_password.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:gravity="center"
     android:orientation="vertical"
-    tools:context=".ui.forgotpassword.ForgotPasswordFragment">
+    tools:context=".ui.auth.forgotpassword.ForgotPasswordFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
