+++ app/src/main/res/navigation/auth_navigation.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/loginFragment"
-        android:name="com.bangkit.alpaca.ui.login.LoginFragment"
+        android:name="com.bangkit.alpaca.ui.auth.login.LoginFragment"
         android:label="fragment_login"
         tools:layout="@layout/fragment_login" >
         <action
@@ -19,7 +19,7 @@
     </fragment>
     <fragment
         android:id="@+id/registrationFragment"
-        android:name="com.bangkit.alpaca.ui.registration.RegistrationFragment"
+        android:name="com.bangkit.alpaca.ui.auth.registration.RegistrationFragment"
         android:label="fragment_registration"
         tools:layout="@layout/fragment_registration" >
         <action
@@ -30,7 +30,7 @@
     </fragment>
     <fragment
         android:id="@+id/forgotPasswordFragment"
-        android:name="com.bangkit.alpaca.ui.forgotpassword.ForgotPasswordFragment"
+        android:name="com.bangkit.alpaca.ui.auth.forgotpassword.ForgotPasswordFragment"
         android:label="fragment_forgot_password"
         tools:layout="@layout/fragment_forgot_password" >
         <action
